KTX2 support updates: ETC1S/BasisLZ, zstd (native), ASTC HDR, updated basis-universal
