+++ Cargo.toml
@@ -159,6 +159,7 @@ default = [
   "smaa_luts",
   "sysinfo_plugin",
   "tonemapping_luts",
+  "zstd_rust",
   "vorbis",
   "webgl2",
   "x11",
@@ -307,6 +308,7 @@ trace = ["bevy_internal/trace", "dep:tracing"]
 
 # Basis Universal compressed texture support
 basis-universal = ["bevy_internal/basis-universal"]
+basis-universal-compressor = ["bevy_internal/basis-universal-compressor"]
 
 # BMP image format support
 bmp = ["bevy_internal/bmp"]
@@ -357,7 +359,9 @@ webp = ["bevy_internal/webp"]
 zlib = ["bevy_internal/zlib"]
 
 # For KTX2 supercompression
-zstd = ["bevy_internal/zstd"]
+zstd = ["bevy_internal/zstd"] # A marker feature indicating zstd support is required for a particular feature. One must select the "zstd_rust" or "zstd_native" feature.
+zstd_rust = ["bevy_internal/zstd_rust"] # Pure-rust zstd implementation (safer)
+zstd_native = ["bevy_internal/zstd_native"] # Binding to zstd C implementation (faster)
 
 # FLAC audio format support
 flac = ["bevy_internal/flac"]
