+++ crates/bevy_image/src/lib.rs
@@ -10,11 +10,20 @@ pub mod prelude {
     };
 }
 
+#[cfg(all(
+    feature = "zstd",
+    not(feature = "zstd_rust"),
+    not(feature = "zstd_native")
+))]
+compile_error!(
+    "There's a Bevy feature requiring zstd decompression support but no zstd backend is selected. Enable the \"zstd_rust\" or \"zstd_native\" feature."
+);
+
 mod image;
 pub use self::image::*;
 #[cfg(feature = "basis-universal")]
 mod basis;
-#[cfg(feature = "basis-universal")]
+#[cfg(feature = "basis-universal-compressor")]
 mod compressed_image_saver;
 #[cfg(feature = "dds")]
 mod dds;
@@ -26,10 +35,14 @@ mod hdr_texture_loader;
 mod image_loader;
 #[cfg(feature = "ktx2")]
 mod ktx2;
+#[cfg(all(feature = "ktx2", feature = "basis-universal"))]
+mod ktx2_using_basisu;
+#[cfg(feature = "ktx2")]
+mod ktx2_using_rust;
 mod texture_atlas;
 mod texture_atlas_builder;
 
-#[cfg(feature = "basis-universal")]
+#[cfg(feature = "basis-universal-compressor")]
 pub use compressed_image_saver::*;
 #[cfg(feature = "dds")]
 pub use dds::*;
@@ -41,6 +54,10 @@ pub use hdr_texture_loader::*;
 pub use image_loader::*;
 #[cfg(feature = "ktx2")]
 pub use ktx2::*;
+#[cfg(all(feature = "ktx2", feature = "basis-universal"))]
+pub use ktx2_using_basisu::*;
+#[cfg(feature = "ktx2")]
+pub use ktx2_using_rust::*;
 pub use texture_atlas::*;
 pub use texture_atlas_builder::*;
