+++ crates/bevy_internal/Cargo.toml
@@ -29,6 +29,7 @@ sysinfo_plugin = ["bevy_diagnostic/sysinfo_plugin"]
 
 # Texture formats that have specific rendering support (HDR enabled by default)
 basis-universal = ["bevy_image/basis-universal", "bevy_render/basis-universal"]
+basis-universal-compressor = ["bevy_image/basis-universal-compressor", "bevy_render/basis-universal-compressor"]
 dds = [
   "bevy_image/dds",
   "bevy_render/dds",
@@ -42,6 +43,8 @@ ktx2 = ["bevy_image/ktx2", "bevy_render/ktx2"]
 # For ktx2 supercompression
 zlib = ["bevy_image/zlib"]
 zstd = ["bevy_image/zstd"]
+zstd_rust = ["bevy_image/zstd_rust"]
+zstd_native = ["bevy_image/zstd_native"]
 
 # Image format support (PNG enabled by default)
 bmp = ["bevy_image/bmp"]
