+++ docs/cargo_features.md
@@ -51,7 +51,8 @@ The default feature set enables most of the expected features of a game engine, 
 |vorbis|OGG/VORBIS audio format support|
 |webgl2|Enable some limitations to be able to use WebGL2. Please refer to the [WebGL2 and WebGPU](https://github.com/bevyengine/bevy/tree/latest/examples#webgl2-and-webgpu) section of the examples README for more information on how to run Wasm builds with WebGPU.|
 |x11|X11 display server support|
-|zstd|For KTX2 supercompression|
+|zstd_native|For KTX2 zstd supercompression using a [zstd](https://crates.io/crates/zstd) binding (faster)|
+|zstd_rust|For KTX2 zstd supercompression using pure-rust [ruzstd](https://crates.io/crates/ruzstd) (safer)|
 
 ### Optional Features
